H2 H3 18px, <p>s and <h2>s spacing(margin and padding)
trying to make the padding/margin above the paragraphs and h2s and images the same height as it is in Xd. This is difficult bc doesn't look the same even though write # dimentions. I also did 0px on other parts to cancel out any extra I didn't want to make size accurate for where I was placing it.